Added grid search and fixed rjmcmc algo
- The RJMCMC sampler now contains methods to do forward and reverse grid
search proposals.
- The new proposals required transformations between the constrained and
unconstrained parameter space so I also added these to a new file
"util.py" (note this is not complete - assumes the parameters are real
variables so things like simplex and covariance matrices won't work).
- Also corrected the RJMCMC acceptance ratio which was missing the ratio
of the continuous proposal probabilities. This is currently returned by
a user-defined function called "continuous_proposal" and could
therefore be made more robust.
- Added a sonar example to demonstrate the new functionality. I had to
modify this example due to an issue with ReddingStan not providing
precision in the logprob values when the base -logprob is large.