Added RJMCMC inside SMC
- Added a new discrete variable type called ReversibleJumpVariable which
stores separate discrete and continuous parameters and samples /
evaluates the proposal using an RJMCMC with the proposal probability set
to the acceptance probability.
- In order not to duplicate code, the RJMCMC algorithm has been
restructured with the addition of separate proposal and evaluate steps.
- Added two new grid search functions in order to evalutate the reverse
proposal probability when we do a discrete move followed by a continuous
move (needed when using NUTS with SMC / comparing across particles).
- Included a new example "sonar_example_smc.py", which demonstrates the
new functionality. To evaluate the proposal probabilities the continuous
proposal functions store the current logprob and other relevant
paramters by piggybacking onto the array of the continuous parameters
(might be worth thinking about other approaches but it will make do for
now!).
- NUTS now has two modes, toggled via the "do_warmup" parameter. If
warmup is enabled (i.e. for RJMCMC) it will run NUTS as normal for an
initial adapatation period when entering a new continuous parameter
space. If disabled (i.e. for SMC), NUTS will continue to adapt the step
size each time a NUTS sample is generated.